refactor: Moved image helpers to image_embedding/utils
