+++ src/image_embedding/mod.rs
@@ -2,6 +2,8 @@ use crate::models::image_embedding::ImageEmbeddingModel;
 const DEFAULT_BATCH_SIZE: usize = 256;
 const DEFAULT_EMBEDDING_MODEL: ImageEmbeddingModel = ImageEmbeddingModel::ClipVitB32;
 
+mod utils;
+
 mod init;
 pub use init::*;
